Rename Lang/StackLang to MathLang (plus spacing)
